Run certain tasks on AWS only
Setting up swap space and staging Docker images is being done only AWS only at the moment so allows those tasks to run conditionally.

The current default variable is a template and raises a warning when run but I can't find a way to conditionally define a variable in a role without using a template.

At the playbook, can do something like this:
```
 vars:
    cm_is_aws: false
 pre_tasks:
   - name: Imply if running targeting AWS
      set_fact:
        cm_is_aws: true
      when: cm_aws_instance_id is defined
```